# src/openapi_server/models/order.py
# coding: utf-8

"""
    Cinema Ticket System API

    API for browsing cinema showings and ordering tickets. An order is created in a PENDING state (seats on hold), filled in step by step with PATCH, and then confirmed. If it is not confirmed before the hold expires, the order lapses on its own and the seats are released - no endpoint is called for that. 

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from datetime import datetime
from pydantic import BaseModel, ConfigDict, Field, StrictFloat, StrictInt, StrictStr, field_validator
from typing import Any, ClassVar, Dict, List, Optional, Union
from openapi_server.models.showing import Showing
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class Order(BaseModel):
    """
    Order
    """ # noqa: E501
    id: Optional[StrictStr] = Field(default=None, json_schema_extra={"examples": ["ord_55871"]})
    status: Optional[StrictStr] = None
    showing_id: Optional[StrictStr] = Field(default=None, alias="showingId")
    showing: Optional[Showing] = None
    seats: Optional[List[StrictStr]] = None
    passenger_name: Optional[StrictStr] = Field(default=None, alias="passengerName")
    contact_email: Optional[StrictStr] = Field(default=None, alias="contactEmail")
    payment_token: Optional[StrictStr] = Field(default=None, description="Token returned by the payment gateway. Never the card number.", alias="paymentToken")
    total_price: Optional[Union[StrictFloat, StrictInt]] = Field(default=None, alias="totalPrice")
    currency: Optional[StrictStr] = Field(default=None, json_schema_extra={"examples": ["ILS"]})
    created_at: Optional[datetime] = Field(default=None, alias="createdAt")
    hold_expires_at: Optional[datetime] = Field(default=None, description="When the hold on the seats expires. If the order is not confirmed by then, the server marks it EXPIRED and releases the seats. No client request is involved. ", alias="holdExpiresAt")
    __properties: ClassVar[List[str]] = ["id", "status", "showingId", "showing", "seats", "passengerName", "contactEmail", "paymentToken", "totalPrice", "currency", "createdAt", "holdExpiresAt"]

    @field_validator('status')
    def status_validate_enum(cls, value):
        """Validates the enum"""
        if value is None:
            return value

        if value not in ('PENDING', 'CONFIRMED', 'CANCELLED', 'EXPIRED',):
            raise ValueError("must be one of enum values ('PENDING', 'CONFIRMED', 'CANCELLED', 'EXPIRED')")
        return value

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of Order from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        # override the default output from pydantic by calling `to_dict()` of showing
        if self.showing:
            _dict['showing'] = self.showing.to_dict()
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of Order from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "id": obj.get("id"),
            "status": obj.get("status"),
            "showingId": obj.get("showingId"),
            "showing": Showing.from_dict(obj.get("showing")) if obj.get("showing") is not None else None,
            "seats": obj.get("seats"),
            "passengerName": obj.get("passengerName"),
            "contactEmail": obj.get("contactEmail"),
            "paymentToken": obj.get("paymentToken"),
            "totalPrice": obj.get("totalPrice"),
            "currency": obj.get("currency"),
            "createdAt": obj.get("createdAt"),
            "holdExpiresAt": obj.get("holdExpiresAt")
        })
        return _obj