# src/openapi_server/models/order_create_request.py
# coding: utf-8

"""
    Cinema Ticket System API

    API for browsing cinema showings and ordering tickets. An order is created in a PENDING state (seats on hold), filled in step by step with PATCH, and then confirmed. If it is not confirmed before the hold expires, the order lapses on its own and the seats are released - no endpoint is called for that. 

    The version of the OpenAPI document: 1.0.0
    Generated by OpenAPI Generator (https://openapi-generator.tech)

    Do not edit the class manually.
"""  # noqa: E501


from __future__ import annotations
import pprint
import re  # noqa: F401
import json




from pydantic import BaseModel, ConfigDict, Field, StrictStr
from typing import Any, ClassVar, Dict, List
try:
    from typing import Self
except ImportError:
    from typing_extensions import Self

class OrderCreateRequest(BaseModel):
    """
    OrderCreateRequest
    """ # noqa: E501
    showing_id: StrictStr = Field(alias="showingId", json_schema_extra={"examples": ["shw_10293"]})
    seats: List[StrictStr] = Field(json_schema_extra={"examples": [["A5", "A6"]]})
    __properties: ClassVar[List[str]] = ["showingId", "seats"]

    model_config = {
        "populate_by_name": True,
        "validate_assignment": True,
        "protected_namespaces": (),
    }


    def to_str(self) -> str:
        """Returns the string representation of the model using alias"""
        return pprint.pformat(self.model_dump(by_alias=True))

    def to_json(self) -> str:
        """Returns the JSON representation of the model using alias"""
        # TODO: pydantic v2: use .model_dump_json(by_alias=True, exclude_unset=True) instead
        return json.dumps(self.to_dict())

    @classmethod
    def from_json(cls, json_str: str) -> Self:
        """Create an instance of OrderCreateRequest from a JSON string"""
        return cls.from_dict(json.loads(json_str))

    def to_dict(self) -> Dict[str, Any]:
        """Return the dictionary representation of the model using alias.

        This has the following differences from calling pydantic's
        `self.model_dump(by_alias=True)`:

        * `None` is only added to the output dict for nullable fields that
          were set at model initialization. Other fields with value `None`
          are ignored.
        """
        _dict = self.model_dump(
            by_alias=True,
            exclude={
            },
            exclude_none=True,
        )
        return _dict

    @classmethod
    def from_dict(cls, obj: Dict) -> Self:
        """Create an instance of OrderCreateRequest from a dict"""
        if obj is None:
            return None

        if not isinstance(obj, dict):
            return cls.model_validate(obj)

        _obj = cls.model_validate({
            "showingId": obj.get("showingId"),
            "seats": obj.get("seats")
        })
        return _obj